Fix search provider meson script
The build is failing without this fix the first time if ninja tries to
build in parallel, because the kolibri-daemon-dbus.h and
shell-search-provider-dbus.h files are not present when it tries to
build the kolibri-gnome-search-provider.

This patch moves the gnome.gdbus_codegen call outside of the library
declaration to be able to add those source files to the executable.