Log in automatically in the Kolibri desktop app
Ask for a login token when the DBus proxy inits.

Add the login token to the conditions for considering the daemon manager started.

Encode the autologin URL inside the "next" query parameter of the
initialize URL. Since the autologin URL also has a "next" query
parameter, this can effectively be chained.

https://phabricator.endlessm.com/T32613