In kolibri-daemon, automatically provision Kolibri
With this change, kolibri-daemon generates an automatic_provision.json
file before starting Kolibri. Because Kolibri must be initialized early
in the application's lifecycle, it would be a nontrivial effort to make
this happen conditionally. For simplicity, kolibri-daemon will always
generate this file, even when Kolibri is running as a session service.
This is a change from the previous release, where automatic provisioning
only applied for Kolibri as a system service.

It is possible to go back to the old behaviour by running kolibri-daemon
with the environment variable
`KOLIBRI_DAEMON_DISABLE_AUTOMATIC_PROVISION=yes`.

https://github.com/learningequality/kolibri-installer-gnome/issues/87
https://github.com/endlessm/endless-key-flatpak/issues/7