metainfo: Use appstreamcli to validate, rather than appstream-util
appstream-util comes from libappstream-glib, which is outdated and
unmaintained. appstreamcli is the more modern replacement, and it comes
from libappstream.

It implements more up to date validation rules, in particular it has
raised the maximum length of screenshot captions from 50 characters to a
more manageable 100 characters.

The downside is that it more strictly validates
`<compulsory_for_desktop>`, and rejects ‘endless’ because it’s not
officially registered as a desktop ID. Work around that for now by
temporarily removing that element from the XML during validation. In the
long term, we need to register the ID:
https://github.com/ximion/appstream/pull/522.

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>

Fixes: https://github.com/endlessm/endless-key-flatpak/issues/34