Remove login token expiry in kolibri-daemon
With the change to use Kolibri's get_os_user hook, we would like for the
user to be able to log in automatically at any time while using Kolibri.
At the moment it is impossible for kolibri-gnome to predict when a token
will be requested and generate a new one, but it is reasonably safe to
reuse an existing token for the duration of a user's session.