Always do automatic provisioning as a system service
In eos-kolibri, the KOLIBRI_APP_AUTOMATIC_PROVISION environment variable
is set for desktop users, but not for the kolibri-daemon system service.
As a result, the daemon application never knows to do automatic
provisioning. To work around this, if kolibri-daemon is running as a
system service we will force automatic provisioning to happen.